Add cache_busting support to document editing
Add a cache_busting boolean option to the Document's to_json and
canonical methods.  If set this will add the Document's updated_at
timestamp in UNIX epoch format onto the end of the page_image_url.
This will prevent browsers from displaying a previously downloaded
document image after it's been edited.

The cache_busting option is only selected by the document viewer's
template if the currently logged in user is allowed to edit the
document.