Add P3P encoded privacy policy
Fix for github issue number: #37.  The addition of a P3P privacy
file and the accompanying http header will force IE version 9 to accept
cookies when DocumentCloud is loaded into an IFrame.

Nginx will also need to have an add_header line added to it's
configuration.  I've made the change in the documentcloud.conf
template file; it will also have to be applied to existing servers.