Backup Solr index.
Create a new CloudCrowd action to snapshot the Solr index, tar/gzip it and store it on S3.

I've been unable to find out if the /solr/replication?command=backup http command:
  * Returns immediately, and you then need to repeatedly request status to determine if it's done
  * Or whether it waits until the snapshot is completed to return the response.
On my system with a very small index, it returns immediately and then status says its completed.  We'll have to run some tests on the larger search index to determine which method we need to accommodate.