Commit document changes manually and raise Solr autoCommit time.
When testing the new solr4 branch, newly uploaded documents where not
appearing in search until after solr was restarted.

This was because commit was not being called on the index.  autoCommit
was enabled, but the openSearcher option wasn't.

This extends the run interval for autoCommit to 10 minutes, and
enables autoSoftCommit every 5 minutes.  It also calls commit manually
on the index after a document is inserted or updated.

Documents will appear in the search immediately after uploading and
the somewhat expensive commit will be called less often.