Upgrade sunspot_solr to the latest pre-release version
This version incorporates a fix for the sunspot:solr:stop rake task
failing to terminate the Java process.  The previous version was
over-escaping the sunspot:solr:start command, causing a sub-process
to be created and incorrect PID logged.  The change we need is
discussed at: https://github.com/sunspot/sunspot/pull/221