Bugfix for account not recording second identity login
On reflection I don't think other code should be able to update the
identity hash on an account directly.  Removed that method and also
the Account.new({profile_incomplete=>true }) bit that got in there
somehow.

As the code-base is right now, Rails throws an exception immediately on
logging in when it attempts to record the organization_id.  I'm
returning nil for right now to continue testing.