Request additional information after a social login.
When the user login originates from omniauth using third party
authentication, show the user what their first and last name will be
displayed as and give them an option to update it.

Still needs additional checks on the account to determine whether or
not to run the account validations, and the input form needs
instructions.