Allow specifying Sanitation level as an option to html_attr
Now that Annotations can be created by the general public, it's
desirable that they are only allowed a restricted subset of html.

This extends the html_attr method of the Sanitized module to use the
Rails convention of passing options as the last values in a
meta-programming class method.  The level can now be specified like so:

  html_attr :sanitized_attr, :another_sanitized_attr, :level=>:basic

The default level remains :super_relaxed as it was before