Add method to set temporarily account's membership
Since the role is no longer directly accessible on the Accounts, it
isn't return by to_json.  For that we need canonical

When an account is updated by an admin, the role is set properly in
accounts_controller, but the returned json doesn't include it and the UI
never updates.

This adds a scoping method to accounts that will temporarily set the
membership to a given organization (if it's valid).  The accounts
methods can then refer to membership and pickup on what membership the
account should be acting under.

If this approach looks good, we can rewrite additional Account methods to
use the @membership instance var.