Make i18n lib a proper class and use namespaces
The older version ran into scope problems when embedding viewer into
workspace.  It would run fine on the un-minimized build, but had 'pack
undefined' errors on staging.

This adds a namespace parameter on the language packs and the lib is
initialized accordingly.

Under the viewer it gets loaded into the DV.t and workspace gets _.t