Rewrite EntityResource to use new query methods
ActiveRecord.all and hash-based options are deprecated, which makes
logic like the original hard to replicate.

It also had a bit of a logic error.  Once the pages method was called,
it'd be cached in the @pages variable and never re-ran even if the query
options were changed.  The excerpts method would also be affected.