Account validations were't running in Admin signup
The email on Account isn't validated unless it belongs to
memberships. We'd done this so that we could support email-less accounts
from facebook & twitter.

The admin controller was creating the account, then adding the
memberships.  This meant that the email validations weren't ran.

Also the account was checking for email with .exists?  This will always
query the database vs .any? which will first see if there are any
in-memory associations, then query if there are not.