Webkit and IE compat for document viewer download
An iframe will always have readyState == 'complete' on Webkit if the
content isn't render-able inside the iframe (like with zip files).

This was causing the iframe to be discarded immediately after it was
created, before the zip had a chance to download.

IE doesn't use readyState == 'complete', it uses 'interactive'