Fix slug generation for long titles without spaces
If a Document title had a long title (>50 characters)
without any spaces, the #ensure_titled method would fail
to extract the words and would set the slug to blank.
The document would then fail the "Slug can't be blank" validation
and fail to save.

The change attempts to first split on spaces, underscores, and dashes.
If that fails, it will just take the first 50 characters of the title.