Increase connection_pool, configure reaper
When multiple workers attempt to join/communicate at once the pool is being exhausted.
There's also no sense in waiting for 5 seconds since we're on a good connection.
If it can't get one immediately, it should give up. Also configure the pool
reaper in case a thread happens to leak a connection.