Don't include CloudCrowd into Kernel name-space
CloudCrowd's constants such as PROCESSING, FAILED, COMPLETE, etc where
polluting the namespace and leading to confusing when the console was used.

This moves only the relevant classes into the Kernel name-space
for easier querying while leaving everything else defined only
under CloudCrowd.

It's a pretty gnarly hack, but is only used in the console, so... yay?