Set RAILS_ENV from ARGV before initialization
The Rails.env defaults to: `ENV["RAILS_ENV"] || "development"`

There are numerous spots in the Rails code-base that cache the value inside strings and they're called before the rake
tasks are.

This means that when a rake task kicks off, the environment is already set and is really impossible to change inside of a
task chain.

This presents a problem since our tasks use the format of "rake <environment> <task>"  The environment task would appear
to succeed, but wouldn't change all aspects to reflect the environment, leaving some things still referring to
development.

This checks to see if the environment is set as the first argument and sets the ENV to reflect it before requiring any
Rails code.