docker: Patch the crap out of DC until it works
Don't even try to count the number of heinous sins committed herein.
You'll be here a while.

Unfortunately DocumentCloud makes all kinds of fairly weird assumptions
about how SSL is terminated in front of it that make running it
*without* SSL (or with SSL terminated by someone else downstream) really
hard.

So we just patched the shit out of it until it worked.

This is probably outright dangerous at this stage. Don't use it.

That said, you can now run DocumentCloud and get to the admin interface
without it blowing up. Which is progress. I think.