Update to latest jQuery file upload
The API for the file upload has changed considerably.  The most notable
changes is that each of the callbacks take the same two parameters:
"event and data"

The event is a standard jQuery event, and the data contains the state
of the file upload.

Callbacks are now reponsible for interrogating the data object in order to
obtain the information they need.

Another significant change is that the uploader now can upload files
in parallel by itself.  This means that the file upload dialog no longer
has to manually start the next file downloading once a file is completed.