Fix upgrading reviewer account to full privileges
The "upgrade_reviewer_to_real" method hadn't been upgraded from when
the account had an "organization" attribute so was throwing
errors when it attempted to write to it.

This no longer updates any old annotations to the new organization.
Our thinking is that since annotations were left by the reviewer
for the inviting organization, they should remain attributed
to that organization.