Streaming & find_each for Accounts CSV generation
The accounts download was taking an long time and large
amount of memory to generate.  This helps, but it's still
running 4 sql queries per account row.  We'll have to start
employing counter_cache or something similar to gain more efficiency.

This extracts the existing CSV streaming code from doc_hits into
a "deliver_csv" method.  We should consider further extracting that
into a mixin if additional controllers need it in the future.