Upload models use state to indicate upload status
Add a "state" flag to the UploadDocument model in an attempt to simplify the
upload logic.  The state defaults to "pending" and is set to "uploading",
and is then set to either "success" or "error".

This enables the upload dialog and it's tiles to listen on the model/collection
for state related events and update themselves accordingly, and the collection to
provide querying functions.