Run validations on email if it's present
Previously the validation was skipped if the account didn't
have any memberships.  This was done so they can be created
from oauth2 providers that do not return an email address.

However, the side effect was that emails were not validated when
the account is first created, since the workflow is to
create an account and then link it to an organization.

Discovered as part of #155