Restore region-specific AWS subdomains
With #195 we allowed installations to override the AWS region, but had to revert (56a6a28f440ee78f7c24ac600a36fd3291c86f7c) the insertion of region into the S3 subdomain because it's not supported for the default region. Now we special-case the asset root to only insert region if it's not the default.