Return correct position of TextScanner match
We made `TextScanner` UTF-safe, but have been storing the end of the match position instead of the beginning. Fix that and fix the test that confirms this.

We now need to reprocess existing documents to regenerate proper entity positions (or run a script that simply adjusts the offsets, since they're wrong by a known amount).

Affects #187