Return DC-rooted image paths for private docs
Prior to this change, `Document#page_image_url` (which, among other things, feeds `resources.thumbnail` on the document JSON) returned an S3-rooted URL for the images of private documents.

With this change, we now supply the correct DC-rooted URLs for private docs, which get routed to `DocumentsController#send_page_image()`.