Use JS for note's pixel ping after all
Sigh. @knowtheory points out that `request.referer` isn't a reliable indicator that an embed page was iframed. It may also have simply been linked-to.

Which means if we want to ensure pixel ping is only generated when this is in an iframe, we need JS.

Effectively reverts e799c25a062a8a3bb223350e7168c4984b799bf1