Don’t pass options to iframe embed wrappers
I stuck these on the iframe wrapper so that if enhance.js wanted, it could be aware of them. But since this is the iframe version, we haven’t mapped the keys to their JS counterparts (see `DC::Embed::Base::Config.dump` and the `map_keys`/`output_as` stuff), so they’re usless to JS-land.

#412